Restructure oracle into cmd folder
